services/federation-service/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
httpx